Chatbot1/requirements.txt
Flask
flask-cors
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
gunicorn
